pyproject.toml
[project]
name = "trabalho3persistencia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beanie>=2.0.1",
    "email-validator>=2.3.0",
    "faker>=40.1.2",
    "fastapi>=0.128.0",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
